Backend/Django/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
chardet==3.0.4
charset-normalizer==2.1.1
Django==3.2.12
django-cors-headers==3.13.0
django-sslserver==0.22
djangorestframework==3.14.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==4.9.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.21.6
pandas==1.3.5
protobuf==3.20.1
python-dateutil==2.8.2
pytrends==4.7.3
pytz==2022.2.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
